# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sponsor-reservation-drafts:
    post:
      operationId: createSponsorReservationDraft
      tags: [checkout]
      summary: Claim one currently open placement briefly and issue a capability cookie
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SponsorReservationDraftInput" }
      responses:
        "201":
          description: Expiring reservation draft created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SponsorReservationDraft" }
        "404": { description: Campaign or spot is unavailable }
        "409": { description: Spot is already being reserved }
        "429": { description: Too many requests }
  /sponsor-reservation-drafts/{draftId}/logo/request-url:
    post:
      operationId: requestSponsorReservationDraftLogoUpload
      tags: [checkout, storage]
      summary: Issue a capability-bound sponsor logo upload intent
      parameters: [{ $ref: "#/components/parameters/ReservationDraftId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SponsorLogoUploadRequest" }
      responses:
        "201":
          description: Single-use logo upload intent and signed URL
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SponsorLogoUploadIntent" }
        "400": { description: Upload metadata is not allowed }
        "404": { description: Draft not found or not authorized }
        "409": { description: Draft expired or unavailable }
  /sponsor-reservation-drafts/{draftId}/logo/{intentId}/finalize:
    post:
      operationId: finalizeSponsorReservationDraftLogoUpload
      tags: [checkout, storage]
      summary: Verify and finalize a sponsor logo upload intent once
      parameters:
        - $ref: "#/components/parameters/ReservationDraftId"
        - name: intentId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        "200":
          description: Logo intent finalized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SponsorLogoUploadIntent" }
        "400": { description: Object metadata does not match the intent }
        "404": { description: Draft or intent not found or not authorized }
        "409": { description: Draft or intent is expired or already used }
...
components:
  securitySchemes:
    operatorBearer:
      type: http
      scheme: bearer
      bearerFormat: API token
  parameters:
...
    ReservationDraftId:
      in: path
      name: draftId
      required: true
      schema: { type: string, minLength: 1, maxLength: 100 }
...
  schemas:
...
    SponsorReservationDraftInput:
      type: object
      additionalProperties: false
      required: [campaignId, spotIndex]
      properties:
        campaignId: { type: string, minLength: 1, maxLength: 100 }
        spotIndex: { type: integer, minimum: 0, maximum: 19 }
    SponsorReservationDraft:
      type: object
      required: [id, campaignId, spotIndex, status, expiresAt]
      properties:
        id: { type: string }
        campaignId: { type: string }
        spotIndex: { type: integer }
        status: { type: string, enum: [issued] }
        expiresAt: { type: string, format: date-time }
    SponsorLogoUploadRequest:
      type: object
      additionalProperties: false
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        size: { type: integer, minimum: 1, maximum: 20000000 }
        contentType: { type: string, enum: [image/svg+xml, application/pdf] }
    SponsorLogoUploadIntent:
      type: object
      required: [id, purpose, objectPath, expectedMimeType, expectedSizeBytes, expectedFileName, status, expiresAt]
      properties:
        id: { type: string }
        purpose: { type: string, const: sponsor_reservation_draft_logo }
        objectPath: { type: string, pattern: "^/objects/uploads/[A-Za-z0-9-]+$" }
        expectedMimeType: { type: string, enum: [image/svg+xml, application/pdf] }
        expectedSizeBytes: { type: integer }
        expectedFileName: { type: string }
        status: { type: string, enum: [issued, finalized, consumed, revoked] }
        expiresAt: { type: string, format: date-time }
        uploadURL: { type: string, format: uri }
...